almost compiling

+++ src/parse/parse_tool.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_tool.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zseignon <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: zseignon <zseignon@student.le-101.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/20 13:26:43 by zseignon          #+#    #+#             */
-/*   Updated: 2020/02/20 13:27:55 by zseignon         ###   ########lyon.fr   */
+/*   Updated: 2020/03/12 12:40:49 by zseignon         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,9 +32,9 @@ static int		tunnelcmp(const char *s1, const char *s2)
 
 	ss1 = 0;
 	ss2 = 0;
-	while (g_white_space[s1[ss1]] == 0)
+	while (g_white_space[(int)s1[ss1]] == 0)
 		ss1 += 1;
-	while (g_white_space[s2[ss2]] == 0)
+	while (g_white_space[(int)s2[ss2]] == 0)
 		ss2 += 1;
 	if (ss1 == ss2)
 		return (ft_memcmp(s1, s2, ss1));
@@ -47,7 +47,7 @@ int				rseek(t_anthill *ah, t_room *rdata, char *p)
 	size_t			n;
 
 	n = 0;
-	while (n < ah->nb_room)
+	while (n < (size_t)ah->nb_room)
 	{
 		if (tunnelcmp(rdata[n].name, p) == 0)
 			return ((int)n);
